dapi friend syncing

+++ DashSync/Models/Entities/DSContactEntity+CoreDataProperties.h
@@ -15,8 +15,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<DSContactEntity *> *)fetchRequest;
 
-@property (nonatomic, assign) uint64_t blockHeight;
-@property (nonatomic, assign) uint64_t documentRevision;
+@property (nonatomic, assign) uint32_t blockHeight;
+@property (nonatomic, assign) uint32_t documentRevision;
 @property (nullable, nonatomic, retain) NSData *encryptionPublicKey;
 @property (nullable, nonatomic, copy) NSString *documentScopeID;
 @property (nullable, nonatomic, copy) NSString *username;
